refactor: Move theme module(s) to `/app/core/theme.ts`
