+++ app/README.md
@@ -1,19 +1,17 @@
-# Front-end Application
+# Web Application (Front-end)
 
 ## Directory Structure
 
 `├──`[`common`](./common) — Common (shared) React components<br>
-`├──`[`core`](./core) — Core modules, React hooks, environment variables, etc.<br>
+`├──`[`core`](./core) — Core modules, React hooks, customized theme, etc.<br>
 `├──`[`dialogs`](./dialogs) — React components implementing modal dialogs<br>
 `├──`[`icons`](./icons) — Custom icon React components<br>
 `├──`[`menus`](./menus) — React components implementing popup menus<br>
 `├──`[`public`](./public) — Static assets such as robots.txt, index.html etc.<br>
 `├──`[`routes`](./routes) — Application routes and page (screen) components<br>
-`├──`[`theme`](./theme) — Application theme - colors, fonts, paddings, etc.<br>
 `├──`[`global.d.ts`](./global.d.ts) — Global TypeScript declarations<br>
-`├──`[`index.html`](./index.html) — HTML markup for single-page app (SPA)<br>
-`├──`[`index.html`](./index.html) — HTML markup for single-page app (SPA)<br>
-`├──`[`index.tsx`](./index.ts) — Application entry point<br>
+`├──`[`index.html`](./index.html) — HTML page containing application entry point<br>
+`├──`[`index.tsx`](./index.tsx) — Single-page application (SPA) entry point<br>
 `├──`[`package.json`](./package.json) — Workspace settings and NPM dependencies<br>
 `├──`[`tsconfig.ts`](./tsconfig.json) — TypeScript configuration<br>
 `└──`[`vite.config.ts`](./vite.config.ts) — JavaScript bundler configuration ([docs](https://vitejs.dev/config/))<br>
